Image_processing&recog/face_recognition_api/requirements.txt
# Deep Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.26.2
facenet-pytorch==2.5.3
Pillow==10.1.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Additional Dependencies
python-dotenv==1.0.0
websockets==11.0.3